feat: tela de Pedidos de Venda (orçamento→confirmação→expedição→faturamento) + campo tipo de Produto
- Angular: nova área vendas/pedidos — listagem (pedidos.ts/html/scss), formulário
  de orçamento com itens em FormArray (pedido-form) e tela de detalhe com ações
  de ciclo de vida (confirmar/recalcular/expedir/faturar/reabrir/cancelar) e
  breakdown de impostos (pedido-detalhe); rotas em app.routes.ts e item de menu
  em web-layout.ts; home.ts aciona "Nova Venda" em vez do placeholder antigo
- Angular: pedido.model.ts ganha campos fiscais (valorTotalNf, valorIbs, valorCbs,
  valorIs, valorIss, valorRetencoes) refletindo o retorno de faturamento
- Angular: produto.model.ts/produtos-form — campo tipo (MERCADORIA/SERVICO) com
  codigoServico e classTrib condicionais no formulário de Produtos
- Angular: <p-toast> único global (removido das telas de pedidos), editar pedido
  agora recarrega do backend em vez de reusar o objeto da listagem, setTimeout
  nos maps de cliente/vendedor pra evitar ExpressionChangedAfterChecked, e
  responsividade dos itens do formulário
- operacoes-service: PedidoResponseDTO/PedidoAssembler passam itens+histórico
  também na resposta de faturamento; PedidoController usa o nome do produto
  (não o UUID) na mensagem de produto inativo; CadastroServiceClient.ProdutoRef
  ganha campo nome
- operacoes-service: fix — PedidoService.atualizarItens faz flush() explícito
  entre o delete e o insert de itens, evitando violar uq_pedido_item_pedido_produto
  quando o Hibernate atrasa o delete derivado pro fim do flush
- fiscal-service: fix — cast ::bigint em tenant_id nas queries de TabelaFiscalJdbc
  (matriz_tributaria/retencao_config), acompanhando a troca de Tenant.id de UUID
  pra bigint; TabelaFiscalJdbcTest atualizado para tenant_id numérico
- liquibase-service: novo changelog fiscal-schema-015 incluído no master
- cadastro-service: fix — EnderecoRepository troca inner join por left join em
  e.estabelecimento, pra não excluir endereços de pessoa física (sem
  estabelecimento) da consulta
- common/Constants.java: corrige comentário do placeholder de PEDIDO_PRODUTO_INATIVO

+++ fiscal-service/src/main/java/com/l/erp/fiscalservice/services/fiscal/TabelaFiscalJdbc.java
@@ -134,9 +134,9 @@ public class TabelaFiscalJdbc implements TabelaFiscal {
                AND uf_origem = :ufOrigem
                AND uf_destino = :ufDestino
                AND ncm_nbs IN (:ncmNbs, :fallback)
-               AND (tenant_id = :tenantId OR tenant_id IS NULL)
+               AND (tenant_id = :tenantId::bigint OR tenant_id IS NULL)
                AND vigente_de <= :data AND (vigente_ate IS NULL OR vigente_ate > :data)
-             ORDER BY CASE WHEN tenant_id = :tenantId THEN 0 ELSE 1 END,
+             ORDER BY CASE WHEN tenant_id = :tenantId::bigint THEN 0 ELSE 1 END,
                       CASE WHEN ncm_nbs = :ncmNbs THEN 0 ELSE 1 END,
                       vigente_de DESC
              LIMIT 1
@@ -148,9 +148,9 @@ public class TabelaFiscalJdbc implements TabelaFiscal {
             SELECT aliquota_pct, valor_minimo_base
               FROM fiscal.retencao_config
              WHERE tributo = :tributo
-               AND (tenant_id = :tenantId OR tenant_id IS NULL)
+               AND (tenant_id = :tenantId::bigint OR tenant_id IS NULL)
                AND ativo = true
-             ORDER BY CASE WHEN tenant_id = :tenantId THEN 0 ELSE 1 END
+             ORDER BY CASE WHEN tenant_id = :tenantId::bigint THEN 0 ELSE 1 END
              LIMIT 1
             """;
